refactor: split aboutUs and login pages
