+++ client/app/aboutUs/_components/StatsSection.js
@@ -1,0 +1,36 @@
+'use client'
+import React from 'react'
+import { Clock, CheckCircle, Users, Headphones } from 'lucide-react'
+import { stats } from '../_utils/constants'
+
+const iconMap = {
+  Clock,
+  CheckCircle,
+  Users,
+  Headphones
+}
+
+export default function StatsSection() {
+  return (
+    <section className="py-16 bg-white">
+      <div className="container mx-auto px-4">
+        <div className="grid md:grid-cols-2 lg:grid-cols-4 gap-8">
+          {stats.map((stat, index) => {
+            const IconComponent = iconMap[stat.icon]
+            return (
+              <div key={index} className="text-center group">
+                <div className="bg-gradient-to-br from-slate-50 to-blue-50 p-8 rounded-2xl border border-slate-100 hover:border-blue-200 transition-all duration-300 hover:shadow-xl">
+                  <div className="w-16 h-16 bg-gradient-to-br from-blue-500 to-indigo-500 rounded-xl flex items-center justify-center mx-auto mb-4 group-hover:scale-110 transition-transform duration-300">
+                    <IconComponent className="h-8 w-8 text-white" />
+                  </div>
+                  <div className="text-3xl font-bold text-slate-800 mb-2">{stat.value}</div>
+                  <div className="text-slate-600 font-medium">{stat.label}</div>
+                </div>
+              </div>
+            )
+          })}
+        </div>
+      </div>
+    </section>
+  )
+}
