+++ client/app/aboutUs/_type/index.d.ts
@@ -1,0 +1,19 @@
+export interface PromiseItem {
+  icon: string
+  title: string
+  description: string
+  gradient: string
+}
+
+export interface TeamMember {
+  role: string
+  message: string
+  icon: string
+  gradient: string
+}
+
+export interface StatItem {
+  value: string
+  label: string
+  icon: string
+}
