+++ client/app/aboutUs/_utils/constants.js
@@ -1,0 +1,54 @@
+export const promises = [
+  {
+    icon: 'FileText',
+    title: '과정의 투명성',
+    description:
+      '신청 과정 하나하나를 실시간으로 투명하게 공개합니다. 서류 접수부터 최종 승인까지 모든 단계를 온라인 대시보드와 알림을 통해 바로 확인하실 수 있습니다.',
+    gradient: 'from-blue-500 to-cyan-500'
+  },
+  {
+    icon: 'Award',
+    title: '결과의 전문성',
+    description:
+      '10년 이상 집중해온 전문가들로 구성되어 있습니다. 가장 최신 법규를 정확히 이해하고, 수많은 성공 사례를 통해 쌓은 노하우로 안정적이고 확실하게 책임집니다.',
+    gradient: 'from-green-500 to-emerald-500'
+  },
+  {
+    icon: 'Heart',
+    title: '소통의 진정성',
+    description:
+      '베트남 현지 생활의 어려움을 이해하는 한국어 소통 전문가가 처음 문의부터 비자 수령까지 모든 과정을 직접 책임지고 함께합니다.',
+    gradient: 'from-orange-500 to-red-500'
+  }
+];
+
+export const team = [
+  {
+    role: '대표 / 비자 전략 총괄',
+    message:
+      '베트남에서 비자 문제로 어려움을 겪는 한국 분들을 보며 "이건 아니다"라는 생각으로 이 일을 시작했습니다. 저희 팀은 오직 고객님의 불안감을 해소하고 신뢰를 드리는 데 모든 초점을 맞추고 있습니다.',
+    icon: 'Target',
+    gradient: 'from-blue-600 to-indigo-600'
+  },
+  {
+    role: '노동허가서 & 거주증 전문 실무팀장',
+    message:
+      '15년 이상의 실무 경험을 통해 베트남 노동허가서와 거주증 발급 과정의 모든 변수와 노하우를 축적했습니다. 가장 복잡하고 까다로운 서류 준비부터 현지 관공서 소통까지 모든 전문성을 동원하겠습니다.',
+    icon: 'Shield',
+    gradient: 'from-green-600 to-emerald-600'
+  },
+  {
+    role: '고객 소통 & 지원 팀',
+    message:
+      '비자 절차는 낯설고 어렵게 느껴질 수 있습니다. 저희 팀은 고객님의 입장에서 모든 질문에 성심껏 답변하고, 진행 상황을 가장 빠르고 정확하게 안내해 드리는 소통의 최전선입니다.',
+    icon: 'Headphones',
+    gradient: 'from-purple-600 to-pink-600'
+  }
+];
+
+export const stats = [
+  { value: '10+', label: '년 전문 경험', icon: 'Clock' },
+  { value: '99.8%', label: '비자 승인률', icon: 'CheckCircle' },
+  { value: '10,000+', label: '성공 사례', icon: 'Users' },
+  { value: '24/7', label: '고객 지원', icon: 'Headphones' }
+];
