+++ client/app/login/_type/index.d.ts
@@ -1,0 +1,6 @@
+export interface LoginInput {
+  email: string
+  password: string
+  rememberMe: boolean
+  autoLogin: boolean
+}
